+++ OptiGUI/src/main/kotlin/opekope2/optigui/internal/BuiltinFilterFactories.kt
@@ -6,14 +6,14 @@ import net.fabricmc.loader.api.VersionParsingException
 import net.minecraft.client.MinecraftClient
 import net.minecraft.resource.ResourceManager
 import net.minecraft.util.Identifier
-import opekope2.optigui.filter.factory.FilterFactoryContext
-import opekope2.optigui.filter.factory.FilterFactoryResult
 import opekope2.optigui.InitializerContext
 import opekope2.optigui.filter.*
 import opekope2.optigui.filter.IFilter.Result.Companion.mismatch
+import opekope2.optigui.filter.factory.FilterFactoryContext
+import opekope2.optigui.filter.factory.FilterFactoryResult
 import opekope2.optigui.interaction.Interaction
 import opekope2.optigui.properties.*
-import opekope2.util.*
+import opekope2.optigui.util.*
 import org.apache.commons.text.StringEscapeUtils.unescapeJava
 import java.time.Month
 import java.time.Month.*
