+++ OptiGUI/src/main/kotlin/opekope2/optigui/util/NumberOrRange.kt
@@ -1,4 +1,4 @@
-package opekope2.util
+package opekope2.optigui.util
 
 import opekope2.optigui.filter.EqualityFilter
 import opekope2.optigui.filter.IFilter
